backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
celery==5.4.0
redis==5.2.1
playwright==1.49.0
playwright-stealth==1.0.6
python-dotenv==1.0.1
httpx==0.28.1
